Update elevator.py
This is a more improved version of the same code. I made sure to space out the inputs and I fixed a few bugs.(For example there was bug where when you gave a incorrect input and then a correct input it would then skip to the part that says "Anything else?") The thing we need to implement is the parallel code idea. Also another bug that needs to be fix is if press close door or open door more than once in succession it should say that the door is already open or closed. I'll probably fix that though.